refactor: move name mangling helpers out of JvmName
