+++ main/src/ca/uwaterloo/flix/language/phase/jvm/JvmName.scala
@@ -17,7 +17,6 @@
 
 package ca.uwaterloo.flix.language.phase.jvm
 
-import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast.SourceLocation
 import ca.uwaterloo.flix.util.InternalCompilerException
 
@@ -55,93 +54,6 @@ object JvmName {
     val descriptor = desc.descriptorString()
     val parts = descriptor.substring(1, descriptor.length - 1).split("/")
     JvmName(parts.init.toList, parts.last)
-  }
-
-  val RootPackage: List[String] = Nil
-
-  /**
-    * Constructs a concatenated string using `JvmName.Delimiter`. The call
-    * `mkClassName("Tuple2", List(Object, Int, String))` would
-    * result in the string `"Tuple2$Obj$Int32$Obj"`.
-    */
-  def mkClassName(prefix: String, args: List[String]): String = {
-    val cPrefix = mangle(prefix)
-    if (args.isEmpty) s"$cPrefix${Flix.Delimiter}"
-    else s"$cPrefix${Flix.Delimiter}${args.map(mangle).mkString(Flix.Delimiter)}"
-  }
-
-  def mkClassName(prefix: String, arg: String): String =
-    mkClassName(prefix, List(arg))
-
-  def mkClassName(prefix: String): String =
-    mkClassName(prefix, Nil)
-
-  /**
-    * Performs name mangling on the given string `s` to avoid issues with special characters.
-    */
-  def mangle(s: String): String = {
-    // Fast path: most names contain no special characters, so we avoid
-    // allocating a new string entirely.
-    if (!containsSpecialChar(s)) s else mangleSlow(s)
-  }
-
-  /**
-    * Returns `true` if `s` contains at least one character that must be mangled.
-    */
-  private def containsSpecialChar(s: String): Boolean = {
-    var i = 0
-    while (i < s.length) {
-      if (mangleReplacement(s.charAt(i)) != null) {
-        return true
-      }
-      i = i + 1
-    }
-    false
-  }
-
-  /**
-    * Mangles `s` in a single pass, replacing every special character with its mangled form.
-    */
-  private def mangleSlow(s: String): String = {
-    val sb = new StringBuilder(s.length + 16)
-    var i = 0
-    while (i < s.length) {
-      val c = s.charAt(i)
-      val replacement = mangleReplacement(c)
-      if (replacement != null) {
-        sb.append(Flix.Delimiter)
-        sb.append(replacement)
-      } else {
-        sb.append(c)
-      }
-      i = i + 1
-    }
-    sb.toString
-  }
-
-  /**
-    * Returns the mangled replacement for the special character `c`, or `null` if `c` is not special.
-    */
-  private def mangleReplacement(c: Char): String = c match {
-    case '+' => "plus"
-    case '-' => "minus"
-    case '*' => "asterisk"
-    case '/' => "fslash"
-    case '\\' => "bslash"
-    case '<' => "less"
-    case '>' => "greater"
-    case '=' => "eq"
-    case '&' => "ampersand"
-    case '|' => "bar"
-    case '^' => "caret"
-    case '~' => "tilde"
-    case '!' => "exclamation"
-    case '#' => "hashtag"
-    case ':' => "colon"
-    case '?' => "question"
-    case '@' => "at"
-    case '.' => "dot"
-    case _ => null
   }
 
   //
@@ -215,9 +127,7 @@ object JvmName {
   // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Flix Names ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   //
 
-  val DevFlixRuntime: List[String] = List("dev", "flix", "runtime")
-
-  val FlixError: JvmName = JvmName(DevFlixRuntime, mkClassName("FlixError"))
+  val FlixError: JvmName = JvmName(Mangle.DevFlixRuntime, Mangle.mkClassName("FlixError"))
 
 }
